// server/prisma/generated/models/Users.ts
export type UsersModel = runtime.Types.Result.DefaultSelection<Prisma.$UsersPayload>
// ...
export type $UsersPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Users"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    email: string
    password: string
  }, ExtArgs["result"]["users"]>
  composites: {}
}